requirements.txt
django==2.2.4
djangorestframework==3.10.3
django-sendgrid-v5==0.8.0
Pillow==6.1.0
